Allow tags to be set in a specific order, for research paper authors
Adding a 'sort' parameter set to true allows the tags to pick up a
sequence number and thereby be set in a specific order on a per-post
basis. This is used for the author list in research publications where
the order is significant.

The majority of this work is by Nick Ogden, completed in late 2013/early
2014.